Fix keying repeater every few seconds
The EchoLink machine was keying the repeater every few seconds. This
was corrected by disabling the "roger" tone in SVXLink. It appears
that the "roger" tone would key the repeater, which would cause
SVXLink to retransmit the repeater's "roger" tone, which in turn
caused SVXLink to emit a "roger" tone, creating an endless loop that
made the repeater essentially useless.

Other changes:

* Set squelch detection to CTCSS
* Correct CTCSS frequency to 151.4 Hz
* Disable short and long idents
* Disable "Repeater" logic - while this machine connects to a repeater,
  it is not itself a repeater